homework1/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
mysql-connector-python==8.0.27
yfinance==0.1.74
protobuf==5.28.1
